Cleaned Code and Fixed Something
Somehow the SerializeField sections were not set and it took a while to realize that. On the other hand, I cleaned the code a little more and changed the order in which plane disabling goes. It now finds the biggest plane, it removes it from the list and then disables the rest. This might be less error prone.