Anchors Basic Implementation + issue solving
I got used with ARAnchors and even implemented them on the main menu button, for the no transition part. Also spent almost 2 hours reinstalling assets, packages and so on and checking with logcat to find out why suddenly LeanTouch does not work. Of course it was a check I forgot to make in PlaneSelection. One if statement sent me on a 2 hour session of frustration and debugging, but it is finally done.