NoTransition/SimplifiedTransition
I finally figured it out. I fixed the buttons appearing before the ARSession could be ready. Then I extracted the model's material so that I can modify its rendering mode in the Inspector. And Finally I researched and used a coroutine to give the fade-in effect to the buttons for the simplifiedTransition. That's the gist of it.