Started work on options menu and debugging.
I decided and made the main buttons that will go on the options menu and started working on implementing them. I got stopped in my tracks though by a few problematic discoveries that made me reimport the buttons assets because of a problem with the misalignment of the axes. I also did fix the simple transitions fade, where the buttons that remained in RenderMode Transparent where see-through on the sides. Also, there were some other small problems that got solved and some other that got created.