Made functional the first option from the options Menu and started working on a JSON
So I changed the direction of the arrows in the option buttons and then implemented the first thing a user can change in the options menu, which is the TransitionType. It works fine though it could probably do with more efficiency. I also started working on a JSON for saving and loading settings and I am thinking about encryption too, though there isn't really sensitive data there yet or might never be. The transparency problem still is unresolved.