Resolved most of the frame-based animations and more
So I updated the logic for SimpleTransition and for LeftRightArrows because they were frame-dependent and the speed at which they ran on the unity game-view was way different from the one on the phone. I also did fix some bugs for the options buttons transitions such as the button moving out of place if the arrows were pressed repeatedly and the name of the button rotating instead of staying in place in certain situations.