Added a "return" type button on the options menu
I added a new button on the options menu that lets you go back to the main menu. I made the model and added it to unity. Afterwards I separated the main menu from the Transitions script because of possible future problems such as the singleton affecting the buttons. Then I found out that because of the ARAnchor the main menu buttons were getting destroyed and so, for now, I have commented that out since it might not actually be needed. I also implemented the logic behind the close/return button, but I need to adjust some other scripts in order to make it all work as intended.