"Back" Function template finished
So I finished implementing the back function of the close button completely and I have adapted everything for it with delegates. It is now ready and could be used in other areas too. I have also changed the material of the close button to a brighter shade of red, the same one as the text on the buttons.