Fixed a problem regarding the "forward" part of a model for all models. Also fixed some small things.
Basically the forward direction from a model in MagickaVoxel is not the same as the one in unity. It is reversed. To be more concise, the z axis, was pointing towards the back of the button instead of the front. All major buttons had this problem and it was affecting ComplexTransitions. I fixed it along with some other small things.