Adjusted some things regarding "Complex Transitions"
I did realise that there were some small errors accumulating because of the rotation of the buttons. Those small differences affected the eulerAngles of the buttons and their position. While I fixed the eulerAngles part, the position one will probably remain like that because it is not noticeable in a normal situation and it resets at each run of the app so it is for the moment not worth the effort.