Polished the switch between menus.
So I made sure to reposition buttons when changes are made to transitionType and I also applied the currentOptions on the visual part of the buttons in the OptionsMenu. So now if go into options, change something and close without saving, when you go back to that menu, it will show the current options and not the changes you made before.