A problem never really stays dead huh
Of course I told solved everyday created some other problem. So I fixed it, it is a rough solution, but it works. I removed the name of the buttons from the button hierarchy, but not there was a problem with the fades because TextMeshPro does not have _color property and so I worked around that.