Made EyePassword script better and finished design for register form
So the EyePassword script was actually made to work for the visibility switch from the login form. Of course I needed this in the register form too, but it wasn't modular enough, so I fixed that and integrated it into the system I had already. I also did set-up the register form from a design standpoint.