Fixed two big problems on mobile
The two problems were unnoticeable on pc and have caused a great deal of trouble. The first one was related on my use of delegates and how I set some data regarding position and rotation for the 2d instantiated Notes. The second problem which took me hours to try different solutions, was related to how the Unity InputField works and how that was a problem for my AR setting. But hey I'm finally free of those, thought it cost me a lot of time.